requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Essential Google AI Services Only
google-generativeai==0.3.2
google-cloud-translate==3.12.1
google-auth==2.23.4

# Basic dependencies
requests==2.31.0
python-dotenv==1.0.0
